requirements.txt
requests==2.31.0
pandas==2.2.1
pprintpp==0.4.0
matplotlib==3.8.3
streamlit==1.32.2
streamlit-card==1.0.0
pillow==10.2.0
CairoSVG==2.7.1